+++ packages/patternfly-4/react-charts/src/components/ChartTheme/ChartTheme.ts
@@ -37,6 +37,8 @@ interface ChartThemeColorInterface {
   gray: string;
   green: string;
   multi: string;
+  multiOrdered: string;
+  multiUnordered: string;
   orange: string;
   purple: string;
 }
@@ -47,6 +49,12 @@ interface ChartThemeVariantInterface {
   light: string;
 }
 
+/**
+ * The multiOrdered theme is intended for ordered charts; donut, pie, bar, & stack
+ * The multiUnordered theme is intended for unordered charts; area & line
+ *
+ * Note: multi defaults to multiOrdered
+ */
 export const ChartThemeColor: ChartThemeColorInterface = {
   blue: 'blue',
   cyan: 'cyan',
@@ -55,6 +63,8 @@ export const ChartThemeColor: ChartThemeColorInterface = {
   gray: 'gray',
   green: 'green',
   multi: 'multi',
+  multiOrdered: 'multi-ordered',
+  multiUnordered: 'multi-unordered',
   orange: 'orange',
   purple: 'purple'
 };