+++ packages/patternfly-4/react-integration/demo-app-ts/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.12.10](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@2.12.9...demo-app-ts@2.12.10) (2019-07-22)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [2.12.9](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@2.12.8...demo-app-ts@2.12.9) (2019-07-22)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [2.12.8](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@2.12.7...demo-app-ts@2.12.8) (2019-07-22)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [2.12.7](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@2.12.6...demo-app-ts@2.12.7) (2019-07-22)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
 ## [2.12.6](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@2.12.5...demo-app-ts@2.12.6) (2019-07-18)
 
 **Note:** Version bump only for package demo-app-ts